% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanBDfile.R
\name{cleanBDfile}
\alias{cleanBDfile}
\title{cleanBDfile}
\usage{
cleanBDfile(path)
}
\arguments{
\item{path}{Specify the path to the xls file containing the Batch Analysis results}
}
\value{
Returns a data frame with columns describing sample information and data
}
\description{
This function reorganizes an output file from BD Accuri C6 Plus software.
}
\examples{
batchAnalysisResults <- dir(system.file("exampleFC.xls"))
cleanBDfile(batchAnalysisResults)
}
